refactor: remove legacy assistant and session_api domains and consolidate to agent domain
