--- src-tauri/src/mcp/builtin/attachments/utils.rs
@@ -1,2 +1,0 @@
-// This module is kept for backward compatibility but contains no active utilities.
-// All response creation is now handled through MCPResult types.
